create separate elements file structure
